+++ rpc/backend/utils.go
@@ -292,7 +292,7 @@ func ShouldIgnoreGasUsed(res *abci.ResponseDeliverTx) bool {
 func GetLogsFromBlockResults(blockRes *tmrpctypes.ResultBlockResults) ([][]*ethtypes.Log, error) {
 	blockLogs := [][]*ethtypes.Log{}
 	for _, txResult := range blockRes.TxsResults {
-		logs, err := evmtypes.DecodeTxLogsFromEvents(txResult.Data, uint64(blockRes.Height))
+		logs, err := evmtypes.DecodeTxLogsFromEvents(txResult.Data, txResult.Events, uint64(blockRes.Height))
 		if err != nil {
 			return nil, err
 		}